Se ajudta errores en codigo

+++ backend032/src/main/java/com/misiontic/backend032/modelos/Agencias.java
@@ -33,7 +33,7 @@ public class Agencias {
     public long getIdAgencia() {
         return idAgencia;
     }
-    public void setIdAgencia(long idAgencia) {
+    public void setIdAgencia(int idAgencia) {
         this.idAgencia = idAgencia;
     }
